Demonstration of Abstract Classes
Following is a demonstration of abstract classes and their applications in OOB. This particular program was assigned class work. I enjoyed working on this program as I was able to incorporate an algorithm I learned on my own time which allows for alternating user input which is stored in arrays. Description is provided in comments withing AccountArray.Java